requirements_qrcode.txt
alembic==1.3.1
APScheduler==3.6.3
astroid==2.3.3
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==3.2
dnspython==1.16.0
docutils==0.15.2
eventlet==0.25.1
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Bcrypt==0.7.1
Flask-Injector==0.12.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-User==1.0.2.2
Flask-WTF==0.14.2
gpio==0.3.0
greenlet==0.4.15
idna==2.8
injector==0.18.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.3.0
mccabe==0.6.1
minimalmodbus==1.0.2
monotonic==1.5
netifaces==0.10.9
passlib==1.7.2
pid==2.2.5
Pillow==8.2.0
psutil==5.6.7
psycopg2==2.8.4
psycopg2-binary==2.8.4
py-cpuinfo==5.0.0
pyasn1==0.4.8
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pyserial==3.4
pysmb==1.2.6
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.11.1
python-socketio==4.4.0
pytz==2019.3
qrcode==6.1
requests==2.22.0
service==0.6.0
setproctitle==1.1.10
six==1.13.0
soupsieve==2.1
SQLAlchemy==1.3.11
typed-ast==1.4.0
typing==3.7.4.1
typing-extensions==3.7.4.1
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
yappi==1.3.2
